+++ riverpie/lib/riverpie.dart
@@ -1,7 +1,7 @@
 export 'package:riverpie/src/async_value.dart';
 export 'package:riverpie/src/container.dart';
 export 'package:riverpie/src/notifier/base_notifier.dart'
-    show ReduxNotifierOverrideExt, Reducer;
+    show ReduxNotifierOverrideExt, Reducer, TestableReduxNotifier;
 export 'package:riverpie/src/notifier/emittable.dart';
 export 'package:riverpie/src/notifier/notifier_event.dart';
 export 'package:riverpie/src/notifier/types/async_notifier.dart';